feat: separate of concerns
